restructured rss feed categories
