+++ app/src/main/res/layout-w900dp/article_list.xml
@@ -10,7 +10,7 @@
     android:divider="?android:attr/dividerHorizontal"
     android:orientation="horizontal"
     android:showDividers="middle"
-    tools:context=".ArticleListActivity">
+    tools:context=".tech.TechArticleListActivity">
 
     <!--
     This layout is a two-pane layout for the Articles
@@ -27,7 +27,7 @@
         android:layout_marginLeft="16dp"
         android:layout_marginRight="16dp"
         app:layoutManager="LinearLayoutManager"
-        tools:context="com.jasminebreedlove.techandsportsnews.ArticleListActivity"
+        tools:context="com.jasminebreedlove.techandsportsnews.tech.TechArticleListActivity"
         tools:listitem="@layout/article_list_content" />
 
     <FrameLayout
