+++ app/src/main/res/layout/activity_article_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ArticleListActivity">
+    tools:context=".tech.TechArticleListActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
