+++ app/src/main/res/layout/activity_sportsarticle_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".SportsArticleDetailActivity"
+    tools:context=".sports.SportsArticleDetailActivity"
     tools:ignore="MergeRootFrame">
 
     <android.support.design.widget.AppBarLayout
