+++ app/src/main/res/layout/activity_sportsarticle_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".SportsArticleListActivity">
+    tools:context=".sports.SportsArticleListActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/sports_app_bar"
