+++ app/src/main/res/layout/article_detail.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_width="match_parent"
     android:orientation="vertical"
-    tools:context=".ArticleDetailFragment">
+    tools:context=".tech.TechArticleDetailFragment">
 
     <TextView
         android:id="@+id/article_pub_date"
