+++ app/src/main/res/layout/article_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".ArticleListActivity"
+    tools:context=".tech.TechArticleListActivity"
     tools:listitem="@layout/article_list_content" />
