+++ app/src/main/res/layout/sportsarticle_detail.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_width="match_parent"
     android:orientation="vertical"
-    tools:context=".ArticleDetailFragment">
+    tools:context=".tech.TechArticleDetailFragment">
 
     <TextView
         android:id="@+id/sports_article_pub_date"
