+++ app/src/main/res/layout/sportsarticle_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".SportsArticleListActivity"
+    tools:context=".sports.SportsArticleListActivity"
     tools:listitem="@layout/sportsarticle_list_content" />
